Update for VVIS CPU fonctional 👍
Preparing GPU patch.
The CPU part (nogpu) is almost the same in the exception of 0.01% difference on my test map. But i will fix it later. Now, i just want to make GPU work. Kernel is good and the process is good, but there is a crash on rewrite and cache portal's datas 👼